* Added configs folder to SConstruct * Fixed MSVC warnings * Released 0.0.3 (win32) to test WinSock

+++ src/network/n_sync.c
@@ -250,7 +250,7 @@ void N_send(int client, const char *format, ...)
                                 goto overflow;
                         break;
                 case 'f':
-                        if (!N_send_float(va_arg(va, double)))
+                        if (!N_send_float((float)va_arg(va, double)))
                                 goto overflow;
                         break;
                 case 's':
